// src/app/interfaces/skripsi.ts
export interface Skripsi {
    id: number;
    idUser: number;
    namaMahasiswa: string;
    nim: string;
    idPeminatan: number;
    namaPeminatan: string;
    judulSkripsi: string;
    idLokasi: number;
    namaLokasi: string;
    idDosen: number;
    namaDosen: string;
    persentaseNilaiD: number;
    persentaseNilaiE: number;
    statusPengajuan: number
}